+++ tests/test_download.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+import geopandas as gpd
 import pytest
 import rasterio as rio
 from omegaconf import DictConfig, OmegaConf
@@ -53,12 +54,13 @@ def paris_timeseries_config(paris_config_path: Path) -> GeefetchConfig:
 
 
 @pytest.fixture
-def paris_config_all_s1_bands_path(
+def paris_config_selected_bands_path(
     raw_paris_config: DictConfig, tmp_path: Path, gee_project_id: str
 ) -> Path:
     raw_paris_config.data_dir = str(tmp_path)
     raw_paris_config.satellite_default.gee.ee_project_id = gee_project_id
     raw_paris_config.s1.selected_bands = ["VV", "VH", "angle"]
+    raw_paris_config.gedi.selected_bands = ["rh95", "rh98"]
 
     conf_path = tmp_path / "config.yaml"
     conf_path.write_text(OmegaConf.to_yaml(raw_paris_config))
@@ -97,9 +99,9 @@ def test_download_s1_overwrite_garbage(paris_config_path: Path):
     assert tif_is_clean(downloaded_tif_path)
 
 
-def test_select_bands_s1(paris_config_all_s1_bands_path: Path):
-    download_s1(paris_config_all_s1_bands_path)
-    conf = load(paris_config_all_s1_bands_path)
+def test_select_bands_s1(paris_config_selected_bands_path: Path):
+    download_s1(paris_config_selected_bands_path)
+    conf = load(paris_config_selected_bands_path)
     downloaded_path = next(iter((Path(conf.data_dir) / "s1").glob("s1*.tif")))
     with rio.open(downloaded_path) as ds:
         assert ds.count == 3
@@ -118,6 +120,14 @@ def test_download_gedi_vector(paris_config_path: Path):
         "gedi_vector",
         "gedi_vector_EPSG2154_650000_6860000.parquet",
     )
+
+
+def test_select_bands_gedi_vector(paris_config_selected_bands_path: Path):
+    download_gedi(paris_config_selected_bands_path, vector=True)
+    conf = load(paris_config_selected_bands_path)
+    downloaded_path = next(iter((Path(conf.data_dir) / "gedi_vector").glob("gedi_*.parquet")))
+    gdf = gpd.read_parquet(downloaded_path)
+    assert gdf.columns.to_list() == ["id", "rh95", "rh98", "geometry"]
 
 
 @pytest.mark.slow